Interaction with Boxes
Object snaps in front of player's camera, however it does not move with the viewpoint just the camera itself. When mouse button is held it follows player when released it stops in place. There are some issues with the ray casting such that it doesn't exactly follow the center of the POV, unsure how to fix it.